pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["dadwalakshay <akshayd00@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "0.7.0"
requests = "2.29.0"
python-dotenv = "1.0.0"
rich = "13.3.5"
django = "4.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "3.2.2"
ipdb = "0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
